DL_Code
The first layer was designed as an embedding layer with the pre-trained word embedding, 40 % of the connections were dropped out directly after the embedding layer to prevent overfitting. We used 4 LSTM layers each with 128 units as hidden layers. The first three layers return a sequence, which enabled us to connect the LSTM layers with each other. After the 4th layer we dropped out 40 % of the connections again. Since the data must be categorized into two categories only, we used a dense layer with two neurons and a sigmoid activation function as output layer. We used the RMSprop optimizer to compile the model. The best results have been achieved when the model was trained with 20 epochs and a batch size of 128. The sequence length has been limited to 100.